Extract current authors footer into own component
