+++ frontend/src/app/features/in-app-notifications/entry/actors-line/in-app-notification-actors-line.component.sass
@@ -1,0 +1,16 @@
+@import "src/assets/sass/helpers"
+
+.op-ian-actors
+  display: grid
+  grid-template-columns: auto 1fr
+  grid-column-gap: 5px
+  align-items: center
+  color: var(--gray-dark)
+
+  &--line
+    vertical-align: baseline
+
+  &--date
+    @include text-shortener
+    max-width: 100%
+    line-height: 16px
