+++ frontend/src/app/features/in-app-notifications/entry/in-app-notification-entry.component.html
@@ -87,31 +87,10 @@
     <div
       class="op-ian-item--bottom-line"
     >
-      <div
-        class="op-ian-item--date"
-        [title]="fixedTime"
-        [textContent]="relativeTime$ | async"
-      ></div>
-      <div>
-        <ng-container *ngFor ="let actor of actors | slice:0:3; let last = last">
-          <span *ngIf="last && actors.length > 1 && actors.length < 4">
-            {{ text.and }}
-          </span>
-          <op-principal
-            class="op-ian-item--actors"
-            [principal]="actor"
-            [hideName]="false"
-            [hideAvatar]="true"
-            [link]="!isMobile()"
-          ></op-principal>
-          <span *ngIf="!last && actors.length > 2" [textContent]="', '"></span>
-        </ng-container>
-        <span
-          *ngIf="actors.length > 3"
-        >
-          {{ text_for_additional_authors(actors.length - 3) }}
-        </span>
-      </div>
+      <op-in-app-notification-actors-line
+        [notification]="notification"
+        [aggregatedNotifications]="aggregatedNotifications"
+      ></op-in-app-notification-actors-line>
     </div>
   </ng-container>
   <ng-template #workPackageLoading>
